web_scraper/requirements.txt
bcrypt==4.2.1
cffi==1.17.1
cryptography==44.0.1
mysql-connector-python==9.2.0
paramiko==3.5.1
pycparser==2.22
PyNaCl==1.5.0
pytz==2025.1
sshtunnel==0.4.0
urllib3==2.3.0
python-dotenv==1.0.1
